refactor(create-board): permissions
