+++ backend/src/modules/schedules/interfaces/services/delete.schedules.service.interface.ts
