+++ database/redis/.env.example
@@ -1,2 +1,1 @@
-REDIS_USER=test
 REDIS_PASSWORD=password123
