+++ docker-compose.dev.yaml
@@ -94,8 +94,6 @@ services:
     image: redis
     command: /bin/sh -c "redis-server --requirepass $$REDIS_PASSWORD"     
     environment:
-      # ALLOW_EMPTY_PASSWORD is recommended only for development.
-      # - ALLOW_EMPTY_PASSWORD=yes
       - REDIS_DISABLE_COMMANDS=FLUSHDB,FLUSHALL
     ports:
       - '6379:6379'
